GET punto residuo by ID (#17)
* GET punto residuo by ID

* PuntoVerde list tests

* PuntoResiduo GET tests

* UTC as date zone

* Use Instant instead of LocalDateTime since 'Local' changes between environments

+++ src/main/java/org/circle8/dto/TipoResiduoDto.java
@@ -4,26 +4,26 @@ import org.circle8.controller.response.TipoResiduoResponse;
 import org.circle8.entity.TipoResiduo;
 
 public class TipoResiduoDto {
-	public int id;
+	public long id;
 	public String nombre;
-	
+
 	public static TipoResiduoDto from(TipoResiduo entity) {
 		var tr = new TipoResiduoDto();
 		tr.id = entity.id;
 		tr.nombre = entity.nombre;
 		return tr;
-	}	
-	
+	}
+
 	public TipoResiduoResponse toResponse() {
 		return new TipoResiduoResponse(id, nombre);
 	}
-	
+
 	public TipoResiduo toEntity() {
 		return TipoResiduo.builder()
 				.id(id)
 				.nombre(nombre)
 				.build();
 	}
-	
-	
+
+
 }
